Stuck with a problem with the color buffer
Error message: GL ERROR :GL_INVALID_OPERATION : glDrawElements: attempt to
access out of range vertices in attribute 1

The myColor attribute is the cause of the problem. The vertex shader seems to
access elements beyond the size of the color buffer.

I tried to put more colors than required, and the problem occurs anyway.

Signed-off-by: Francis Giraldeau <francis.giraldeau@gmail.com>